+++ macros/feats/planarWarrior.js
@@ -74,7 +74,7 @@ if (args[0].tag === "OnUse" && args[0].macroPass === "preItemRoll") {
     origin: macroData.itemUuid, //flag the effect as associated to the item used
     disabled: false,
     duration: foundry.utils.duplicate(macroData.item.effects[0].duration),
-    icon: macroData.item.img,
+    img: macroData.item.img,
     label: macroData.item.name,
     name: macroData.item.name,
   };