+++ macros/feats/slayersPrey.js
@@ -24,7 +24,7 @@ if (args[0].tag === "OnUse") {
         ],
         origin: args[0].actorUuid,
         disabled: false,
-        icon: args[0].item.img,
+        img: args[0].item.img,
         label: "Slayer's Prey Hit",
         name: "Slayer's Prey Hit",
       };