+++ macros/spells/hex.js
@@ -29,7 +29,7 @@ if (args[0].tag === "OnUse") {
     origin: args[0].uuid, //flag the effect as associated to the spell being cast
     disabled: false,
     duration: args[0].item.effects[0].duration,
-    icon: args[0].item.img,
+    img: args[0].item.img,
     label: args[0].item.name,
     name: args[0].item.name,
   };