+++ macros/spells/spiritShroud.js
@@ -51,7 +51,7 @@ if (args[0].tag === "OnUse") {
     origin: args[0].itemUuid,
     disabled: false,
     duration: args[0].item.effects[0].duration,
-    icon: args[0].item.img,
+    img: args[0].item.img,
     label: args[0].item.name,
     name: args[0].item.name,
   };